Added Makefile and neated sources some more
